requirements.txt
notion-client==2.2.1
yfinance==0.2.28
requests==2.31.0
python-dotenv==1.0.0
